Move constants and functions, format code
