+++ metrics/src/config.py
@@ -60,3 +60,6 @@ OFFCHAIN_KEY = '__offchain'
 
 DB_CONNECTION_RETRIES = 30
 DB_CONNECTION_INTERVAL = 2
+
+TRANSACTION_COUNT_FIELD = 'transactions_count'
+BACKFILL_DB_DAYS = 30
